remove previous odk mock bindings
