+++ frontend/src/app/components/odk/odk.components.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from "@angular/core";
 import { MaterialComponentsModule } from "../material.components";
 import { OdkTableRowsComponent } from "./odkTableRows/odkTableRows";
 import { CommonModule } from "@angular/common";
-import { ODKFormPopup } from "./odkFormPopup/odkFormPopup";
 import { ODKDesignerIframeComponent } from "./odk.designerIframe";
 import { ODKRecordsPendingComponent } from "./odk.recordsPending";
 
@@ -13,9 +12,9 @@ const customComponents = [
 ];
 
 @NgModule({
-  entryComponents: [ODKFormPopup],
-  declarations: [customComponents, ODKFormPopup],
+  entryComponents: [],
+  declarations: [customComponents],
   imports: [MaterialComponentsModule, CommonModule],
-  exports: [customComponents, ODKFormPopup],
+  exports: [customComponents],
 })
 export class ODKComponentsModule {}
