+++ frontend/src/app/services/odk/altMockImplementation/odkCommon.ts
